Group footer and mobile menu improvements
Adds a footer to the single group view. Also restyles much of the
mobile menu.

Note that the group footer has to be in the loop… so the sidebar also
has to be in the loop… so we cannot use any widgets that also use group
loops in that sidebar.